+++ PSFramework/changelog.md
@@ -1,4 +1,19 @@
 ﻿# CHANGELOG
+## 1.0.32 : 2019-08-11
+ - New: Validation Attribute: PsfValidateTrustedData - equivalent to ValidateTrustedData, but exists on PS3+ (no effect before 5+)
+ - New: Command Import-PSFPowerShellDataFile - wraps around Import-PowerShellDataFile and makes it available on PSv3+
+ - Upd: Parameter Class: Computer : Add support for output of Get-ADDomainController
+ - Upd: ConvertTo-PSFHashtable : Reimplemented as Cmdlet for better performance
+ - Upd: ConvertTo-PSFHashtable : Adding -Inherit parameter, causing the command to pick up missing includes from variables.
+ - Upd: Select-PSFObject : Parameter `-Property` now validates for trusted data
+ - Upd: Tab Completion: PSFramework-Input-ObjectProperty - will now properly unroll arrays to provide completion for the first value in one.
+ - Upd: Register-PSFTeppScriptblock : Changed some internal behavior
+ - Fix: Write-PSFMessage fails with error on localized string when specifying an error record
+ - Fix: Write-PSFMessage fails with error when specifying $null for format values
+ - Fix: Remove-PSFConfig fails to log deleted configuration name
+ - Fix: Register-PSFTaskEngineTask fails to reset correctly
+ - Fix: PsfValidateSet fails unexpectedly under certain circumstances
+
 ## 1.0.19 : 2019-05-21
  - Upd: Import-PSFConfig adding -PassThru parameter.
  - Upd: Write-PSFMessageProxy adding parameters to better support all common redirection scenarios.